+++ cvars-macros/tests/progress.rs
@@ -12,6 +12,7 @@ fn tests() {
     t.pass("tests/derive/test_typed_setters.rs");
     t.pass("tests/derive/test_string_getters.rs");
     t.pass("tests/derive/test_string_setters.rs");
+    t.pass("tests/derive/test_bool.rs");
     t.pass("tests/derive/test_cvar_count.rs");
     t.pass("tests/derive/test_dyn.rs");
     t.pass("tests/derive/test_redefined.rs");
@@ -23,6 +24,7 @@ fn tests() {
     t.pass("tests/fnlike/test_typed_setters.rs");
     t.pass("tests/fnlike/test_string_getters.rs");
     t.pass("tests/fnlike/test_string_setters.rs");
+    t.pass("tests/fnlike/test_bool.rs");
     t.pass("tests/fnlike/test_cvar_count.rs");
     t.pass("tests/fnlike/test_dyn.rs");
     t.pass("tests/fnlike/test_redefined.rs");